// src/app/shared/product-type.ts
type IReview = {
  user?: string;
  name: string;
  email: string;
  rating: number;
  review: string;
  date: string;
}

export interface IProduct {
  id: string;
  sku: string;
  img: string;
  title: string;
  name?: string;
  slug: string;
  unit: string;
  imageURLs: {
    color?: {
      name: string;
      clrCode: string;
      colorID?: number;
    };
    img: string;
  }[];
  parent: string;
  children: string;
  price: number;
  discount: number;
  quantity: number;
  brand: { name: string };
  category: { id?: string; name: string };
  status: string;
  reviews?: IReview[];
  productType: string;
  description: string;
  orderQuantity?: number;
  additionalInformation: { key: string; value: string }[];
  featured?: boolean;
  sellCount: number;
  offerDate?: { startDate: string; endDate: string };
  tags?: string[];
  videoId?: string;
  badge?: string;
  sizes?: {
    name: string;
    colorID: number;
    quantity: number;
  }[];
  sizesByColor?: {
    [colorID: number]: {
      name: string;
      colorID: number;
      quantity: number;

    }[];
  };
}